Renamed test to match conventions
