+++ test/src/support_data/CMakeLists.txt
@@ -8,5 +8,5 @@ OSSIM_SETUP_APPLICATION(ossim-srtm-support-data-test INSTALL COMMAND_LINE COMPON
 OSSIM_SETUP_APPLICATION(ossim-tiff-info-test INSTALL COMMAND_LINE COMPONENT_NAME ossim SOURCE_FILES ossim-tiff-info-test.cpp)
 OSSIM_SETUP_APPLICATION(ossim-wavelength-test INSTALL COMMAND_LINE COMPONENT_NAME ossim SOURCE_FILES ossim-wavelength-test.cpp)
 
-OSSIM_SETUP_APPLICATION(CSEXRB-test INSTALL COMMAND_LINE COMPONENT_NAME ossim SOURCE_FILES CSEXRB-test.cpp)
+OSSIM_SETUP_APPLICATION(CSEXRB-test INSTALL COMMAND_LINE COMPONENT_NAME ossim SOURCE_FILES ossim-nitf-generic-tag-test.cpp)
